mkpart: Allow negative start value when FS-TYPE is not given
The manual had long documented that negative values are
allowed for both start and end values, but until now negative
start values were rejected if FS-TYPE was not given.

Example:

  # parted --script -a optimal /dev/loop0 -- mklabel gpt mkpart primary ext4 -5MiB 100%
  (succeeds)

  # parted --script -a optimal /dev/loop0 -- mklabel gpt mkpart primary -5MiB 100%
  parted: invalid token: -5MiB
  Error: Expecting a file system type.

This commit fixes the latter error.
The issue was an insufficient lookahead in command line parsing,
looking only for digits when skipping over FS-TYPE.
The fix is including the minus '-' in the lookahead.

Originally reported as Debian bug #880035:

  "parted: fails to use negative start value for 'mkpart' command without specyfying FS-TYPE"
  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=880035

Signed-off-by: Niklas Hambüchen <mail@nh2.me>